% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{compute_S_centroid}
\alias{compute_S_centroid}
\title{Computes the conditioning set for centroid linkage hierarchical clustering}
\usage{
compute_S_centroid(X, hcl, K, k1, k2)
}
\arguments{
\item{X}{the n x q data set}

\item{hcl}{hclust object obtained by clustering X}

\item{K}{number of clusters}

\item{k1}{the index of first cluster involved in the test}

\item{k2}{the index of second cluster involved in the test}
}
\value{
Returns an "Intervals" object containing the conditioning set.
}
\description{
Computes the conditioning set for centroid linkage hierarchical clustering
}
\keyword{internal}
