% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{compute_S_ward_gencov}
\alias{compute_S_ward_gencov}
\title{Computes the conditioning set S for ward linkage hierarchical clustering,
w/o assuming isotropic covariance matrix}
\usage{
compute_S_ward_gencov(X, hcl, K, k1, k2, stat)
}
\arguments{
\item{X}{the n x q data set}

\item{hcl}{hclust object obtained by clustering X}

\item{K}{number of clusters}

\item{k1}{the index of first cluster involved in the test}

\item{k2}{the index of second cluster involved in the test}

\item{stat}{the test statistic, \eqn{||\Sigma^{-1/2} x^T \nu||_2}}
}
\value{
Returns an "Intervals" object containing the conditioning set.
}
\description{
Computes the conditioning set S for ward linkage hierarchical clustering,
w/o assuming isotropic covariance matrix
}
\keyword{internal}
