% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{finiteE}
\alias{finiteE}
\title{Make endpoints of intervals finite}
\usage{
finiteE(E)
}
\arguments{
\item{E}{an "Intervals" object or a matrix where rows represents
a union of intervals with \emph{positive} but possibly infinite endpoints.}
}
\value{
This function returns an "Intervals" object or a matrix depending on the input.
}
\description{
This function modifies a union of intervals with positive but possibly infinite endpoints
   into a union of intervals with positive and \emph{finite} endpoints, while ensuring
   the probability of a \eqn{N(0, 1)} falling into it numerically the same.
}
\keyword{internal}
