% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_integer_between_a_b}
\alias{is_integer_between_a_b}
\title{Checks if input is an integer between a and b}
\usage{
is_integer_between_a_b(x, a, b)
}
\arguments{
\item{x}{input to check}

\item{a}{lower}

\item{b}{upper}
}
\value{
Returns TRUE if input is an integer between a and b, FALSE otherwise
}
\description{
Checks if input is an integer between a and b
}
\keyword{internal}
