% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{norm_vec}
\alias{norm_vec}
\title{Takes the l2-norm of a vector.}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{the vector to be normed}
}
\value{
Returns the l2-norm of x.
}
\description{
Takes the l2-norm of a vector.
}
\keyword{internal}
