% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{same_cl}
\alias{same_cl}
\title{Checks if two clusterings are the same up to permutation}
\usage{
same_cl(cl1, cl2, K)
}
\arguments{
\item{cl1}{the first clustering}

\item{cl2}{the second clustering}

\item{K}{the number of clusters}
}
\value{
Returns TRUE if they are the same, and FALSE otherwise
}
\description{
Checks if two clusterings are the same up to permutation
}
\keyword{internal}
