% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R
\name{solve_one_ineq}
\alias{solve_one_ineq}
\title{Solve the roots of quadratic polynomials related to testing for a difference in means}
\usage{
solve_one_ineq(A, B, C, tol = 1e-10)
}
\arguments{
\item{A, }{B, C the coefficients of the quadratic equation.}

\item{tol}{if \eqn{|a|}, \eqn{|b|}, or \eqn{|c|} is not larger than tol, then treat it as zero.}
}
\value{
Returns an "Intervals" object containing NA or the complement of the solution set.
}
\description{
Solves \eqn{ax^2 + bx + c \ge 0}, then returns the complement of the solution set
wrt to the real line, unless the complement is empty, in which case
the function returns NA.
}
\keyword{internal}
