% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{TChisqRatioApprox}
\alias{TChisqRatioApprox}
\title{Approximation of the ratio of two chi-squared probabilities}
\usage{
TChisqRatioApprox(df, E1, E2)
}
\arguments{
\item{df}{degree of freedom of the chi-squared random variable.}

\item{E1, E2}{"Intervals" objects or matrices where rows represents
a union of intervals with \emph{positive and finite} endpoints.}
}
\value{
This function returns the value of the approximation.
}
\description{
This function returns an approximation of \eqn{P(X \in E1)/P(X \in E2)}, where
    \eqn{X} is a central chi-squared random variable with \code{df} degrees of freedom.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.

Canal, Luisa. "A normal approximation for the chi-square distribution."
    Computational statistics & data analysis 48.4 (2005): 803-808.
}
\keyword{internal}
