% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{TNProb}
\alias{TNProb}
\title{Probability of a standard normal in a union of intervals}
\usage{
TNProb(E)
}
\arguments{
\item{E}{an "Intervals" object or a matrix where rows represents
a union of disjoint intervals.}
}
\value{
This function returns the desired probability.
}
\description{
This function returns \eqn{P(Z \in E)}, where \eqn{Z ~ N(0, 1)}.
}
\keyword{internal}
