% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{TNProbEachInt}
\alias{TNProbEachInt}
\title{Probability of a standard normal in a single interval}
\usage{
TNProbEachInt(lo, up)
}
\arguments{
\item{lo, up}{quantiles.}
}
\value{
This function returns the desired probability.
}
\description{
This function returns \eqn{P(lo \le Z \le up)}, where \eqn{Z ~ N(0, 1)}.
}
\keyword{internal}
