% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{TNRatioApprox}
\alias{TNRatioApprox}
\title{Approximation of the ratio of two normal probabilities}
\usage{
TNRatioApprox(E1, E2, scale = NULL)
}
\arguments{
\item{E1, E2}{"Intervals" objects or matrices where rows represents
a union of intervals with \emph{positive and finite} endpoints.}

\item{scale}{scaling parameter.}
}
\value{
This function returns the value of the approximation.
}
\description{
This function returns an approximation of \eqn{P(Z \in E1)/P(Z \in E2)}, where \eqn{Z ~ N(0, 1)}.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.
}
\keyword{internal}
